+++ resources/views/bible/page.blade.php
@@ -17,9 +17,10 @@
     ?>
     <div class="container-fluid">
         <div class="app">
-
+            <search-component></search-component>
             <index-component 
             :books="books"></index-component>
+            
 
             <page-component
                 api-token="{{ $apiToken }}"